Refactor and add unit tests for XCUITest uploads
